feat(daw-standalone): loop_source, fade shapes, master section, RPP track envelopes
Closes four playback-fidelity gaps against the RPP:

- LOOP 1 items tile their source past its end instead of going silent
  (298 items in the showcase session loop).
- Item fades honor REAPER's curve family (fast-start/end quadratic/
  quartic, slow-S smoothstep/smootherstep) instead of always-linear.
- Master section applies: MASTER_VOLUME gain + pan (balance law —
  centre is unity on an already-mixed bus) and MASTERMUTESOLO mute.
  dawfile parse fixed: MASTER_VOLUME fields are floats, not i32s.
- Track envelopes load from the RPP: dawfile's Envelope::from_block
  was a TODO stub returning empty points — now parses EGUID/ACT/VIS/
  ARM/DEFSHAPE/PT/POOLEDENVINST; MUTEENV/WIDTHENV/AUX*ENV block names
  classify as envelopes. Loader converts to renderer conventions
  (pan -1..1 -> 0..1, mute inverted: RPP >0.5 = play) and respects
  ACT 0 as automation-off.

RenderSnapshot now carries the master section alongside the track
graph in the same revision-keyed cache.

+++ features/backends/reaper/dawfile-reaper/src/types/project.rs
@@ -330,11 +330,12 @@ pub struct ProjectProperties {
     pub master_track_view: Option<MasterTrackView>, // MASTERTRACKVIEW 0 0.6667 0.5 0.5 0 0 0 0 0 0 0 0 0 0
     pub master_hw_out: MasterHwOut,                 // MASTERHWOUT 0 0 1 0 0 0 0 -1
     pub master_nch: Option<(i32, i32)>,             // MASTER_NCH 2 2
-    pub master_volume: Option<(i32, i32, i32, i32, i32)>, // MASTER_VOLUME 1 0 -1 -1 1
-    pub master_pan_mode: Option<i32>,               // MASTER_PANMODE 3
-    pub master_pan_law_flags: Option<i32>,          // MASTER_PANLAWFLAGS 3
-    pub master_fx: Option<i32>,                     // MASTER_FX 1
-    pub master_sel: Option<i32>,                    // MASTER_SEL 0
+    /// `MASTER_VOLUME vol pan ? ? ?` — vol is linear gain, pan −1…1.
+    pub master_volume: Option<(f64, f64, f64, f64, f64)>,
+    pub master_pan_mode: Option<i32>,      // MASTER_PANMODE 3
+    pub master_pan_law_flags: Option<i32>, // MASTER_PANLAWFLAGS 3
+    pub master_fx: Option<i32>,            // MASTER_FX 1
+    pub master_sel: Option<i32>,           // MASTER_SEL 0
 
     // Global automation
     pub global_auto: Option<i32>, // GLOBAL_AUTO -1
@@ -1166,8 +1167,7 @@ impl ProjectProperties {
                         tokens[3].as_number(),
                         tokens[4].as_number(),
                     ) {
-                        self.master_volume =
-                            Some((a as i32, b as i32, c as i32, d as i32, e as i32));
+                        self.master_volume = Some((a, b, c, d, e));
                     }
                 }
             }
@@ -1770,7 +1770,7 @@ mod tests {
         assert_eq!(props.master_mute_solo, Some(0));
         assert_eq!(props.master_hw_out, Some((0, 0, 1, 0, 0, 0, 0, -1)));
         assert_eq!(props.master_nch, Some((2, 2)));
-        assert_eq!(props.master_volume, Some((1, 0, -1, -1, 1)));
+        assert_eq!(props.master_volume, Some((1.0, 0.0, -1.0, -1.0, 1.0)));
         assert_eq!(props.master_pan_mode, Some(3));
         assert_eq!(props.master_pan_law_flags, Some(3));
         assert_eq!(props.master_fx, Some(1));
